move non-in-memory-mmr specific code to the top-level mmr crate
